revert: remove unit test

+++ test/unit-test/utest_helpers.h
@@ -38,7 +38,7 @@
  */
 CborError createOtaStreamingMessage( uint8_t * pMessageBuffer,
                                      size_t messageBufferSize,
-                                     int64_t blockIndex,
+                                     int blockIndex,
                                      uint8_t * pBlockPayload,
                                      size_t blockPayloadSize,
                                      size_t * pEncodedSize,
